+++ src/cqfantasy/house/__init__.py
@@ -1,5 +1,6 @@
 from .Body import Body
 from .Roof import Roof
+from .TileGenerator import TileGenerator
 from .House import House
 from .TudorWall import tudor_wall
 from .StuccoBrickBody import StuccoBrickBody